+++ sickbeard/providers/hdspace.py
@@ -15,6 +15,10 @@
 # You should have received a copy of the GNU General Public License
 # along with SickGear.  If not, see <http://www.gnu.org/licenses/>.
 
+try:
+    from collections import OrderedDict
+except ImportError:
+    from requests.compat import OrderedDict
 import re
 import traceback
 
@@ -42,12 +46,15 @@ class HDSpaceProvider(generic.TorrentProvider):
 
         self.url = self.urls['config_provider_home_uri']
 
-        self.username, self.password, self.freeleech, self.minseed, self.minleech = 5 * [None]
+        self.filter = []
+        self.may_filter = OrderedDict([
+            ('f0', ('not marked', False, '')), ('f25', ('FL', True, 'gold|sf')), ('f50', ('F/L', True, 'silver|sf'))])
+        self.username, self.password, self.minseed, self.minleech = 4 * [None]
 
     def _authorised(self, **kwargs):
 
         return super(HDSpaceProvider, self)._authorised(
-            post_params={'uid': self.username, 'pwd': self.password, 'form_tmpl': 'name=[\'"]login[\'"]'})
+            post_params={'uid': self.username, 'form_tmpl': 'name=[\'"]login[\'"]'})
 
     def _search_provider(self, search_params, **kwargs):
 
@@ -57,8 +64,17 @@ class HDSpaceProvider(generic.TorrentProvider):
 
         items = {'Cache': [], 'Season': [], 'Episode': [], 'Propers': []}
 
-        rc = dict((k, re.compile('(?i)' + v)) for (k, v) in {'info': 'torrent-details', 'get': 'download', 'fl': 'free',
-                                                             'peers': 'page=peers', 'nodots': '[\.\s]+'}.items())
+        rc = dict((k, re.compile('(?i)' + v)) for (k, v) in {
+            'info': 'torrent-details', 'get': 'download', 'peers': 'page=peers', 'nodots': '[\.\s]+'}.items())
+        log = ''
+        if self.filter:
+            non_marked = 'f0' in self.filter
+            # if search_any, use unselected to exclude, else use selected to keep
+            filters = ([f for f in self.may_filter if f in self.filter],
+                       [f for f in self.may_filter if f not in self.filter])[non_marked]
+            rc['filter'] = re.compile('(?i)(%s).png' % '|'.join(
+                [self.may_filter[f][2] for f in filters if self.may_filter[f][1]]))
+            log = '%sing (%s) ' % (('keep', 'skipp')[non_marked], ', '.join([self.may_filter[f][0] for f in filters]))
         for mode in search_params.keys():
             for search_string in search_params[mode]:
                 search_string = isinstance(search_string, unicode) and unidecode(search_string) or search_string
@@ -83,7 +99,10 @@ class HDSpaceProvider(generic.TorrentProvider):
                             raise generic.HaltParseException
 
                         for tr in torrent_rows[1:]:
-                            if tr.find('td', class_='header'):
+                            if (tr.find('td', class_='header')
+                                or (any(self.filter)
+                                    and ((non_marked and tr.find('img', src=rc['filter']))
+                                         or (not non_marked and not tr.find('img', src=rc['filter']))))):
                                 continue
                             downlink = tr.find('a', href=rc['get'])
                             if None is downlink:
@@ -91,8 +110,7 @@ class HDSpaceProvider(generic.TorrentProvider):
                             try:
                                 seeders, leechers = [tryInt(x.get_text().strip())
                                                      for x in tr.find_all('a', href=rc['peers'])]
-                                if self._peers_fail(mode, seeders, leechers)\
-                                        or self.freeleech and None is tr.find('img', title=rc['fl']):
+                                if self._peers_fail(mode, seeders, leechers):
                                     continue
 
                                 info = tr.find('a', href=rc['info'])
@@ -109,7 +127,7 @@ class HDSpaceProvider(generic.TorrentProvider):
                     pass
                 except (StandardError, Exception):
                     logger.log(u'Failed to parse. Traceback: %s' % traceback.format_exc(), logger.ERROR)
-                self._log_search(mode, len(items[mode]) - cnt, search_url)
+                self._log_search(mode, len(items[mode]) - cnt, log + search_url)
 
             results = self._sort_seeding(mode, results + items[mode])